requirements.txt
tensorflow-gpu<1.15